requirements.txt
requests>=2.25.1
PyYAML>=5.4.1
pytest>=6.2.2
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.0
